+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 WebSocket:
 
 * Add is_upgrade() free function
+* Document websocket::stream thread safety
 
 API Changes: